requirements.txt
discord.py==1.5.1
redis==3.5.3
PyNaCl==1.4.0
youtube-dl==2020.12.29